+++ spec/directions/bake_further_research_spec.rb
@@ -80,7 +80,7 @@ RSpec.describe Kitchen::Directions::BakeFurtherResearch do
               </div>
               <section class="further-research" data-element-type="further-research">
                 <a href="#first">
-                  <h3 data-type="document-title" id="first_copy_1" itemprop="name">
+                  <h3 data-type="document-title" id="first_copy_1">
                     <span class="os-number">1.1</span>
                     <span class="os-divider"> </span>
                     <span class="os-text" data-type="" itemprop="">First Title</span>
@@ -90,7 +90,7 @@ RSpec.describe Kitchen::Directions::BakeFurtherResearch do
               </section>
               <section class="further-research" data-element-type="further-research">
                 <a href="#second">
-                  <h3 data-type="document-title" id="second_copy_1" itemprop="name">
+                  <h3 data-type="document-title" id="second_copy_1">
                     <span class="os-number">1.2</span>
                     <span class="os-divider"> </span>
                     <span class="os-text" data-type="" itemprop="">Second Title</span>